% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{Plot of sensitivity function\eqn{\Psi}{Psi} for optimal design}
\usage{
plot3D(x, ...)
}
\arguments{
\item{x}{an object of type "\code{tpopt2d}".}

\item{...}{further graphical parameters.}
}
\description{
This function plots the sensitivity function \eqn{\Psi(\boldsymbol x,\xi)} for resulting approximation \eqn{\xi^{**}} of the \eqn{T_{\mathrm{P}}}-optimal design achieved with the help of \code{tpopt2d}. The definition of \eqn{\Psi(\boldsymbol x,\xi)} can be found in the "details" section of function's \code{tpopt2d} specifications.
}
\details{
We examine the shape of the function \eqn{\Psi(x,\xi^{**})} to assess whether the algorithm has converged. If convergence has been achieved, the support points of \eqn{\xi^{**}} (shown as dots) should lie near the local maxima of this function.
Moreover, all local maxima of  \eqn{\Psi(x,\xi^{**})} should attain the same value. If this is not the case, the procedure has likely failed to converge, and the algorithm should be restarted with different parameters.
}
\seealso{
\code{\link{tpopt2d}},
\code{\link{summary_tpopt2d}}
}
