% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodd2D-package.R
\docType{data}
\name{rodd2D-package}
\alias{rodd2D-package}
\title{Optimal Discriminating Designs for Two-Dimensional Predictor Variables}
\description{
This package computes efficient numerical optimal discriminating designs
for two dimensional predictor variables in both the original and the log-transformed models (see the models in Yousefi and Müller, 2023).
It extends the functionality of the \code{rodd} package to the two-dimensional setting.
}
\details{
\code{rodd2D} is a package for the construction of \eqn{T_{\mathrm{P}}}-optimal designs
between a given list of models in both the original and the log-transformed models where predictor variable (\eqn{\boldsymbol x}) is two dimensional.
This approach is based on the algorithms that were developed by Dette et al.(2015).
}
\references{
Atkinson A.C., Fedorov V.V.(1975) The design of experiments for discriminating between two rival models. Biometrika, Volume 62, Issue 1, Pages 57-70.

Atkinson A.C., Fedorov V.V. (1975) Optimal design : Experiments for discriminating between several models. Biometrika, Volume 62, Issue 2, Pages 289-303.

Atkinson, A. C. (2012) Optimum Experimental Designs for Choosing Between Competitive and Non Competitive Models of Enzyme Inhibition. Communications in Statistics - Theory and Methods, 41(13-14), 2283-2296.

Atkinson A.C., Bogacka B.(2014) Optimum designs for the equality of parameters in enzyme inhibition kinetic models. Journal of Statistical Planning and Inference,Volume 144, Pages 47-54,ISSN 0378-3758.

Dette H., Melas V.B., Guchenko R.(2015) BayesianT-optimal discriminating designs. Ann. Statist. 43 (5) 1959 - 1985.

Yousefi, E., Müller, W.G.(2023) Impact of the Error Structure on the Design and Analysis of Enzyme Kinetic Models. Stat Biosci 15, 31-56

Guchenko, R. (2016). \emph{rodd: Optimal discriminating designs}. R package, version 0.2-1.
Available at \url{https://cran.r-project.org/package=rodd}.
}
\author{
Abolfazl Joukar, Elham Yousefi
}
\keyword{datasets}
