% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tpopt2d.R
\name{summary_tpopt2d}
\alias{summary_tpopt2d}
\title{Detailed information about the input}
\usage{
summary_tpopt2d(x, ...)
}
\arguments{
\item{x}{an object of type "\code{tpopt2d}".}

\item{...}{further arguments.}
}
\description{
This function prints detailed information about the input object of class "\code{tpopt2d}".
}
\details{
Call, list of models, list of fixed parameters, resulting design, efficiency by iteration and overall execution time are displayed.
}
\seealso{
\code{\link{tpopt2d}},
\code{\link{plot3D}}
}
