% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpopt2d.R
\name{tpopt2d}
\alias{tpopt2d}
\title{Calculation of optimal discriminating design for two-dimensional predictor variables}
\usage{
tpopt2d(x, w, eta, theta.fix, theta.var, p, x.lower, x.upper, opt)
}
\arguments{
\item{x}{a \eqn{r\times 2} numeric matrix specifying design from \eqn{\mathcal{X}} for initial design, where \eqn{\mathcal{X}} represents the rectangular design region from \eqn{\mathbf{R^2}}.}

\item{w}{a numeric vector indicating weights for the initial design. This vector should have the same length as the number of rows in the initial design. Furthermore, the weights of the design should sum to 1. If this vector is not specified, then the weights are presumed to be equal.}

\item{eta}{a list of models between which suggested optimization should be performed. Every function from this list should be defined in the form of \eqn{\eta_i(\boldsymbol x,\theta_i)}, where \eqn{x} is two dimensional variables from rectangular design region \eqn{\mathcal{X}} and \eqn{\theta_i} is a vector of corresponding model parameters. We will refer to length of this list as \eqn{\nu}.}

\item{theta.fix}{a list of fixed model parameters \eqn{\overline{\theta}_{i}} from the functional \eqn{T_{\mathrm{P}}}. The length of this list should be same as the list of models.}

\item{theta.var}{an array with two dimensions specifying initial values for parameter vectors \eqn{\theta_{i,j}}. The default value here is NULL, which means that initial guess is calculated automatically.}

\item{p}{a \eqn{\nu\times\nu} square table (R-matrix) containing non-negative weights for comparison. The diagonal values of this table should all be zeros. If one want to include comparison of the \eqn{i}'th model with fixed parameters against \eqn{j}'th model with variable parameters into optimization, then he/she should place non-negative weight \eqn{p_{i,j}} into the table; otherwise this weight should be zero.}

\item{x.lower}{a pair lower bound for two dimensional variables \eqn{\boldsymbol x}.If it is not specified, then minimal value of each column from input matrix x is taken.}

\item{x.upper}{a pair upper bound for two dimensional variables \eqn{\boldsymbol x}.If it is not specified, then maximal value of each column from input matrix x is taken.}

\item{opt}{a list of options containing such named fields:
\describe{
\item{x_dis}{a list of points to be used as initial values for two dimensional variables \eqn{\boldsymbol x} when searching for local maxima of the sensitivity function with nloptr package. The x_dis argument is a list that contains x_dis1, x_dis2, and x_dis3. If x_dis3 is "seq", x_dis1 and x_dis2 should be a set of points or sequence. If x_dis3 is "length", x_dis1 and x_dis2 should be the lengths of a sequence. If x_dis3 is "by", x_dis1 and x_dis2 should be increments of the sequence. The default value here is NULL, which means that points are considered automatically.}
\item{method}{a variable indicating the method to be used in inner weight optimization step. The value "1" specify for quadratic programming based procedure and "2" specify for gradient method. See Dette et al.(2015) for more details on that methods.}
\item{max.iter}{maximum number of iterations for the main loop. Reaching this number of iterations is one of the possible stopping conditions.}
\item{des.eff}{desired efficiency for resulted approximation of optimal design. Reaching efficiency of more than des.eff is another stopping condition (to be exact, efficiency lower bound is calculated on each iteration of the algorithm instead of efficiency). See details section for exact definition of efficiency.}
\item{theta.lower}{a list of lower bound for model parameters \eqn{\eta_i(\boldsymbol x,\theta_i)}.This list should have the same length as the list of models.The default value here is NULL, which means that the lower bound is -Inf.}
\item{theta.upper}{a list of upper bound for model parameters \eqn{\eta_i(\boldsymbol x,\theta_i)}.This list should have the same length as the list of models.The default value here is NULL, which means that the upper bound is Inf.}
\item{optim.method}{the method to be used for optimization for the parameters. Method "L-BFGS-B" is optim function with "L-BFGS-B" algorithm. method "optimize" is optimize function for one dimensional optimization and method "BBoptim" is BBoptim function. If it is NULL, optim function by default algorithm will be used.}
\item{support.epsilon}{a value that is used for support points exclusion, if corresponding weights value is less then support.epsilon.}
\item{weights.evaluation.epsilon}{a value that is used in the inner loop for weights evaluation.}
\item{weights.evaluation.max.iter}{ a value that is used in the inner loop for weights evaluation.}
}}
}
\value{
an object of the class "tpopt2d", which is a list inducing following components:
\item{optimal.design}{an \eqn{r\times 3} numeric matrix represents the resulting approximation of the \eqn{T_{\mathrm{P}}}-optimal design. The third row of this matrix contains the weights for the resulting approximation, and the values in this row sum to 1.}
\item{efficiency}{the numeric vector containing efficiency lower bound values by iteration. See details section for definition.}
\item{eta}{the list of models which is identical to the one in the arguments list.}
\item{theta.fix}{the list of fixed model parameters which is exactly the same as one from the arguments list. It remains unchanged in the results.}
\item{theta.var}{a two-dimensional array specifying computed values of the parameter vectors \eqn{\theta_{i,j}} corresponding to the resulting design.}
\item{p, x.lower, x.upper}{comparison table and lower and upper bounds for two dimensional variables \eqn{\boldsymbol x} same as in input.}
\item{max.iter}{max.iter of iterations specified in the options list.}
\item{done.iter}{number of iterations done.}
\item{des.eff}{desired efficiency value specified in the options list.}
\item{time}{overall execution time.}
}
\description{
This function calculates an approximation \eqn{\xi^{**}} of the \eqn{T_{\mathrm{P}}}-optimal design \eqn{\xi^*} for discrimination between a given list of models \eqn{\{\eta_i(\boldsymbol x,\theta_i),\; i = 1,\dots,\nu\}} in both the original and the log-transformed models where predictor variable (\eqn{\boldsymbol x}) is two dimensional, ie \eqn{\boldsymbol x \in \mathfrak{X} \subseteq \mathbf{R^2}}. This approach is based on the algorithms that were developed by Dette et al.(2015). \eqn{T_{\mathrm{P}}}-optimality criterion is defind as:
\deqn{T_{\mathrm{P}}(\xi) = \sum_{i,j=1}^{\nu} p_{i,j} \inf_{\theta_{i,j} \in \Theta_j} \int_{\mathcal{X}} \Big[ \eta_i(\boldsymbol x,\overline{\theta}_{i}) - \eta_j(\boldsymbol x,\theta_{i,j}) \Big]^2 \xi(d\boldsymbol x),}{T_P(xi) = sum_ij p_ij inf_{theta_ij \in Theta_j} int_X [ eta_i(\boldsymbol x,theta_i^) - \eta_j(\boldsymbol x,theta_ij) ]^2 xi(d\boldsymbol x),}
where \eqn{\xi} is an arbitrary design on \eqn{\mathcal{X}} (\eqn{\mathcal{X}} represents the rectangular design region from \eqn{\mathbf{R^2}}), \eqn{\mathrm{P} = \{ p_{i,j} \}_{i,j = 1}^{\nu}} is a table of non-negative weights with zeros on the diagonal (comparison table) and \eqn{\overline{\theta}_{i}} are predefined fixed parameters.
}
\details{
The sensitive function \eqn{\Psi} is defined as
\deqn{\Psi(\boldsymbol x,\xi) = \sum_{i,j=1}^{\nu} p_{i,j} \Big[ \eta_i(\boldsymbol x,\overline{\theta}_{i}) - \eta_j(\boldsymbol x,\widehat{\theta}_{i,j}) \Big]^2, \widehat{\theta}_{i,j} = \arg\inf_{\theta_{i,j} \in \Theta_j} \int_{\mathcal{X}} \Big[ \eta_i(\boldsymbol x,\overline{\theta}_i) - \eta_j(\boldsymbol x, \theta_{i,j}) \Big]^2 \xi(dx).}
The simplified algorithm schema is as follows:
Let \eqn{\xi_s} represents the design obtained on the s'th iteration of the algorithm. Then
\describe{
\item{Step 1.}{Support of the new design \eqn{\xi_{s+1}} consists of all local maximums of function \eqn{\Psi(\boldsymbol x,\xi_s)} on \eqn{\mathcal{X}} united with the support of current design \eqn{\xi_s}.}
\item{Step 2.}{Weights of the new design \eqn{\xi_{s+1}} are calculated so that the functional \eqn{T_{\mathrm{P}}(\xi)} achieves its maximum in the class of all designs with support from previous step.}
}
}
\examples{
\dontrun{
#----------------------------------------------------
#           Example 1--original models
#----------------------------------------------------
#List of models
eta.1 <- function(x1,x2,theta.1){
#noncometitive model
(theta.1[1]*x1) / ((theta.1[2]+x1)*(1+(x2/theta.1[3])))
}
eta.2 <- function(x1,x2,theta.2){
#cometitive model
(theta.2[1]*x1) / ((theta.2[2]*(1+(x2/theta.2[3])))+x1)
}
eta <- list(eta.1, eta.2)
#-----------------------------------------
#initial design
x1 <- c(0,30,0,30,15)
x2 <- c(0,0,40,40,20)
xstart <- cbind(x1,x2)
#List of fixed parameters
theta.1 <- c(10,4.36,5.16)
theta.2 <- c(10,4.36,2.58)
theta.fix <- list(theta.1, theta.2)
theta.1.lower <- rep(0.001,3)
theta.2.lower <- rep(0.001,3)
theta.lower <- list(theta.1.lower, theta.2.lower)
#-----------------------------------------
p1=matrix(c(0,1,0, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(method=1,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)

#gradient method
res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(method=2,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)
#-----------------------------------------
p2=matrix(c(0,0,1, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res2<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p2,x.lower=NULL,x.upper=NULL,
                opt=list(method=1,theta.lower=theta.lower))
plot3D(res2)
summary_tpopt2d(res2)

#gradient method
res2<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p2,x.lower=NULL,x.upper=NULL,
                opt=list(method=2,theta.lower=theta.lower))
plot3D(res2)
summary_tpopt2d(res2)
#-----------------------------------------
p3=matrix(c(0,0.5,0.5, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res3<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p3,x.lower=NULL,x.upper=NULL,
                opt=list(method=1,theta.lower=theta.lower))
plot3D(res3)
summary_tpopt2d(res3)

#gradient method
res3<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p3,x.lower=NULL,x.upper=NULL,
                opt=list(method=2,theta.lower=theta.lower))
plot3D(res3)
summary_tpopt2d(res3)
#--------------------------------------------------------
# An example that shows the algorithm is not sensitive
# to the initial values when searching for local maxima.
#--------------------------------------------------------
p1=matrix(c(0,1,0, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=31,x_dis2=41,x_dis3="length"),
                method=1,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)

res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=125,x_dis2=270,x_dis3="length"),
                method=1,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)
#--------------------------------------------------------
#gradient method
res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=31,x_dis2=41,x_dis3="length"),
                method=2,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)

res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=NULL,
               p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=125,x_dis2=270,x_dis3="length"),
                method=2,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)
#----------------------------------------------------
#           Example 2--log-transformed models
#----------------------------------------------------
#' #List of models
eta.1 <- function(x1,x2,theta.1){
#noncometitive model
log((theta.1[1]*x1) / ((theta.1[2]+x1)*(1+(x2/theta.1[3]))))
}
eta.2 <- function(x1,x2,theta.2){
#cometitive model
log((theta.2[1]*x1) / ((theta.2[2]*(1+(x2/theta.2[3])))+x1))
}
eta <- list(eta.1, eta.2)
#-----------------------------------------
#initial design
eps <- 0.02
x1 <- c(eps,30,eps,30,15)
x2 <- c(0,0,60,60,30)
xstart <- cbind(x1,x2)
#List of fixed parameters
theta.1 <- c(14.3482, 13.8632, 8.7160)
theta.2 <- c(6.0645, 3.2799, 3.3153)
theta.fix <- list(theta.1, theta.2)
theta.1.lower <- rep(eps,3)
theta.2.lower <- rep(eps,3)
theta.lower <- list(theta.1.lower, theta.2.lower)
theta.var1<-array(list(NULL), c(length(eta),length(eta)))
theta.var1[[1,1]]=theta.1
theta.var1[[1,2]]=theta.2
theta.var1[[2,1]]=theta.1
theta.var1[[2,2]]=theta.2
#-----------------------------------------
p1=matrix(c(0,1,0, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=theta.var1,
                p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=c(eps,seq(1,30,length=30)),
                x_dis2=seq(0,60,length=61),x_dis3="seq"),
                method=1,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)

#gradient method
res1<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=theta.var1,
                p=p1,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=c(eps,seq(1,30,length=30)),
                x_dis2=seq(0,60,length=61),x_dis3="seq"),
                method=2,theta.lower=theta.lower))
plot3D(res1)
summary_tpopt2d(res1)
#-----------------------------------------
p2=matrix(c(0,0,1, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res2<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=theta.var1,
                p=p2,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=c(eps,seq(1,30,length=30)),
                x_dis2=seq(0,60,length=61),x_dis3="seq"),
                method=1,theta.lower=theta.lower))
plot3D(res2)
summary_tpopt2d(res2)

#gradient method
res2<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=theta.var1,
                p=p2,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=c(eps,seq(1,30,length=30)),
                x_dis2=seq(0,60,length=61),x_dis3="seq"),
                method=2,theta.lower=theta.lower))
plot3D(res2)
summary_tpopt2d(res2)
#-----------------------------------------
p3=matrix(c(0,0.5,0.5, 0), c(2, 2), byrow = TRUE)

#quadratic programming method
res3<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=theta.var1,
                p=p3,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=c(eps,seq(1,30,length=30)),
                x_dis2=seq(0,60,length=61),x_dis3="seq"),
                method=1,theta.lower=theta.lower))
plot3D(res3)
summary_tpopt2d(res3)

#gradient method
res3<- tpopt2d(x=xstart,w=NULL,eta=eta,theta.fix=theta.fix,theta.var=theta.var1,
                p=p3,x.lower=NULL,x.upper=NULL,
                opt=list(x_dis=list(x_dis1=c(eps,seq(1,30,length=30)),
                x_dis2=seq(0,60,length=61),x_dis3="seq"),
                method=2,theta.lower=theta.lower))
plot3D(res3)
summary_tpopt2d(res3)
}

}
\references{
Atkinson A.C., Fedorov V.V.(1975) The design of experiments for discriminating between two rival models. Biometrika, Volume 62, Issue 1, Pages 57-70.

Atkinson A.C., Fedorov V.V. (1975) Optimal design : Experiments for discriminating between several models. Biometrika, Volume 62, Issue 2, Pages 289-303.

Atkinson, A. C. (2012) Optimum Experimental Designs for Choosing Between Competitive and Non Competitive Models of Enzyme Inhibition. Communications in Statistics - Theory and Methods, 41(13-14), 2283-2296.

Atkinson A.C., Bogacka B.(2014) Optimum designs for the equality of parameters in enzyme inhibition kinetic models. Journal of Statistical Planning and Inference,Volume 144, Pages 47-54,ISSN 0378-3758.

Dette H., Melas V.B., Guchenko R.(2015) BayesianT-optimal discriminating designs. Ann. Statist. 43 (5) 1959 - 1985.

Yousefi, E., Müller, W.G.(2023) Impact of the Error Structure on the Design and Analysis of Enzyme Kinetic Models. Stat Biosci 15, 31-56.
}
\seealso{
\code{\link{plot3D}},
\code{\link{summary_tpopt2d}}
}
